+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.26.2
 ### Changed
 - Simplify cache block import code.
+- Improve cache import performance.
 
 ## 0.26.1
 ### Fixed